+++ specimen/src/org/labkey/specimen/SpecimenManager.java
@@ -28,8 +28,8 @@ import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.specimen.model.SpecimenComment;
-import org.labkey.api.specimen.model.SpecimenTypeSummary;
-import org.labkey.api.specimen.model.SpecimenTypeSummaryRow;
+import org.labkey.specimen.model.SpecimenTypeSummary;
+import org.labkey.specimen.model.SpecimenTypeSummaryRow;
 import org.labkey.api.study.Cohort;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
