+++ specimen/src/org/labkey/specimen/model/SpecimenTypeSummaryRow.java
@@ -13,12 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.labkey.api.specimen.model;
+package org.labkey.specimen.model;
 
-/**
- * User: brittp
- * Created: Jan 31, 2008 11:34:47 AM
- */
 public class SpecimenTypeSummaryRow
 {
     private String _primaryType;
